# docs/swagger.yaml
basePath: /v1
definitions:
  api.ErrorDetail:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorDetail'
      metadata:
        $ref: '#/definitions/api.ResponseMetadata'
      success:
        type: boolean
    type: object
...
  api.ResponseMetadata:
    properties:
      timestamp:
        type: string
      trace_id:
        type: string
    type: object
...
  services.TransactionQueryResponse:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/services.TransactionResponse'
        type: array
    type: object
  services.TransactionResponse:
    properties:
      amount:
        type: number
      closing_balance:
        type: number
      direction:
        type: string
      id:
        type: string
      info:
        additionalProperties: true
        type: object
      price:
        type: number
      quote_symbol:
        type: string
      reason:
        type: string
      sub_account_id:
        type: string
      timestamp:
        type: string
      trading_id:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@tiris.ai
    name: Tiris API Support
    url: https://tiris.ai/support
  description: A Go-based microservice for quantitative trading data management, providing
    RESTful APIs for user management, trading integration, and trading operations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://tiris.ai/terms
  title: Tiris Backend API
  version: "1.0"
paths:
...
  /admin/transactions:
    get:
      description: Lists all transactions with filtering and pagination (admin only)
      parameters:
      - description: Filter by direction
        enum:
        - debit
        - credit
        in: query
        name: direction
        type: string
      - description: Filter by reason
        in: query
        name: reason
        type: string
      - description: Start date (RFC3339 format)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339 format)
        in: query
        name: end_date
        type: string
      - description: Minimum amount filter
        in: query
        name: min_amount
        type: number
      - description: Maximum amount filter
        in: query
        name: max_amount
        type: number
      - default: 100
        description: Number of transactions to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of transactions to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TransactionQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all transactions
      tags:
      - Transactions
  /admin/transactions/{id}:
    get:
      description: Retrieves a transaction by ID (admin only)
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - Transactions
...
  /transactions:
    get:
      description: Retrieves transaction history for the authenticated user with filtering
        and pagination
      parameters:
      - description: Filter by direction
        enum:
        - debit
        - credit
        in: query
        name: direction
        type: string
      - description: Filter by reason
        in: query
        name: reason
        type: string
      - description: Start date (RFC3339 format)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339 format)
        in: query
        name: end_date
        type: string
      - description: Minimum amount filter
        in: query
        name: min_amount
        type: number
      - description: Maximum amount filter
        in: query
        name: max_amount
        type: number
      - default: 100
        description: Number of transactions to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of transactions to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TransactionQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user transactions
      tags:
      - Transactions
  /transactions/{id}:
    get:
      description: Retrieves a specific transaction by ID (must belong to authenticated
        user)
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - Transactions
  /transactions/sub-account/{sub_account_id}:
    get:
      description: Retrieves transaction history for a specific sub-account (must
        belong to authenticated user)
      parameters:
      - description: Sub-account ID
        in: path
        name: sub_account_id
        required: true
        type: string
      - description: Filter by direction
        enum:
        - debit
        - credit
        in: query
        name: direction
        type: string
      - description: Filter by reason
        in: query
        name: reason
        type: string
      - description: Start date (RFC3339 format)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339 format)
        in: query
        name: end_date
        type: string
      - description: Minimum amount filter
        in: query
        name: min_amount
        type: number
      - description: Maximum amount filter
        in: query
        name: max_amount
        type: number
      - default: 100
        description: Number of transactions to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of transactions to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TransactionQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get sub-account transactions
      tags:
      - Transactions
  /transactions/time-range:
    get:
      description: Retrieves transactions within a specific time range for the authenticated
        user
      parameters:
      - description: Start time (RFC3339 format)
        in: query
        name: start_time
        required: true
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: end_time
        required: true
        type: string
      - description: Filter by direction
        enum:
        - debit
        - credit
        in: query
        name: direction
        type: string
      - description: Filter by reason
        in: query
        name: reason
        type: string
      - description: Minimum amount filter
        in: query
        name: min_amount
        type: number
      - description: Maximum amount filter
        in: query
        name: max_amount
        type: number
      - default: 100
        description: Number of transactions to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of transactions to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TransactionQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get transactions by time range
      tags:
      - Transactions
  /transactions/trading/{trading_id}:
    get:
      description: Retrieves transaction history for a specific trading (must
        belong to authenticated user)
      parameters:
      - description: Trading ID
        in: path
        name: trading_id
        required: true
        type: string
      - description: Filter by direction
        enum:
        - debit
        - credit
        in: query
        name: direction
        type: string
      - description: Filter by reason
        in: query
        name: reason
        type: string
      - description: Start date (RFC3339 format)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339 format)
        in: query
        name: end_date
        type: string
      - description: Minimum amount filter
        in: query
        name: min_amount
        type: number
      - description: Maximum amount filter
        in: query
        name: max_amount
        type: number
      - default: 100
        description: Number of transactions to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of transactions to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TransactionQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get trading transactions
      tags:
      - Transactions
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT Bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...